+++ qase-testcafe/src/options.ts
@@ -1,8 +1,5 @@
 // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
 export type ReporterOptionsType = {
-  screenshotsFolder?: string;
-  videosFolder?: string;
-  uploadDelay?: number;
   browser?: {
     addAsParameter?: boolean;
     parameterName?: string;